changed micro to milliseconds in boot time

+++ fun_test/web/fun_test/metrics_models.py
@@ -1206,13 +1206,13 @@ class BootTimePerformance(models.Model):
     interpolated = models.BooleanField(default=False)
     status = models.CharField(max_length=30, verbose_name="Status", default=RESULTS["PASSED"])
     input_date_time = models.DateTimeField(verbose_name="Date", default=datetime.now)
-    output_firmware_boot_time = models.IntegerField(verbose_name="Firmware" ,default=-1)
-    output_flash_type_boot_time = models.IntegerField(verbose_name="Flash type detection" ,default=-1)
-    output_eeprom_boot_time = models.IntegerField(verbose_name="EEPROM Loading", default=-1)
-    output_sbus_boot_time = models.IntegerField(verbose_name="SBUS Loading", default=-1)
-    output_host_boot_time = models.IntegerField(verbose_name="Host BOOT", default=-1)
-    output_main_loop_boot_time = models.IntegerField(verbose_name="Main Loop", default=-1)
-    output_boot_success_boot_time = models.IntegerField(verbose_name="Boot success", default=-1)
+    output_firmware_boot_time = models.FloatField(verbose_name="Firmware" ,default=-1)
+    output_flash_type_boot_time = models.FloatField(verbose_name="Flash type detection" ,default=-1)
+    output_eeprom_boot_time = models.FloatField(verbose_name="EEPROM Loading", default=-1)
+    output_sbus_boot_time = models.FloatField(verbose_name="SBUS Loading", default=-1)
+    output_host_boot_time = models.FloatField(verbose_name="Host BOOT", default=-1)
+    output_main_loop_boot_time = models.FloatField(verbose_name="Main Loop", default=-1)
+    output_boot_success_boot_time = models.FloatField(verbose_name="Boot success", default=-1)
     tag = "analytics"
 
     def __str__(self):
